ADD: Refactor target encoding structure
