+++ src/encoding/target_encoders/__init__.py
@@ -1,0 +1,1 @@
+from .classification_encoder import ClassificationEncoder
